move fixed point constants to fixed point package
