+++ fixedpoint/check.go
@@ -19,8 +19,44 @@
 package fixedpoint
 
 import (
+	"math"
 	"math/big"
 )
+
+const Fix64Scale uint = 8
+const Fix64Factor = 100_000_000
+
+// Fix64
+
+const Fix64TypeMinInt = math.MinInt64 / Fix64Factor
+const Fix64TypeMaxInt = math.MaxInt64 / Fix64Factor
+
+var Fix64TypeMinIntBig = new(big.Int).SetInt64(Fix64TypeMinInt)
+var Fix64TypeMaxIntBig = new(big.Int).SetInt64(Fix64TypeMaxInt)
+
+const Fix64TypeMinFractional = math.MinInt64 % Fix64Factor
+const Fix64TypeMaxFractional = math.MaxInt64 % Fix64Factor
+
+var Fix64TypeMinFractionalBig = new(big.Int).SetInt64(Fix64TypeMinFractional)
+var Fix64TypeMaxFractionalBig = new(big.Int).SetInt64(Fix64TypeMaxFractional)
+
+// UFix64
+
+const UFix64TypeMinInt = 0
+const UFix64TypeMaxInt = math.MaxUint64 / uint64(Fix64Factor)
+
+var UFix64TypeMinIntBig = new(big.Int).SetUint64(UFix64TypeMinInt)
+var UFix64TypeMaxIntBig = new(big.Int).SetUint64(UFix64TypeMaxInt)
+
+const UFix64TypeMinFractional = 0
+const UFix64TypeMaxFractional = math.MaxUint64 % uint64(Fix64Factor)
+
+var UFix64TypeMinFractionalBig = new(big.Int).SetUint64(UFix64TypeMinFractional)
+var UFix64TypeMaxFractionalBig = new(big.Int).SetUint64(UFix64TypeMaxFractional)
+
+func init() {
+	Fix64TypeMinFractionalBig.Abs(Fix64TypeMinFractionalBig)
+}
 
 func CheckRange(
 	negative bool,
